# docs/openapi.json
{
    "components": {
        "schemas": {
            "api.ItemRequest": {
                "properties": {
                    "description": {
                        "example": "A small widget",
                        "maxLength": 2000,
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "name": {
                        "example": "Widget",
                        "maxLength": 200,
                        "type": "string"
                    }
                },
                "required": [
                    "name"
                ],
                "type": "object"
            },
            "api.ItemResponse": {
                "properties": {
                    "description": {
                        "example": "A small widget",
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "id": {
                        "example": 1,
                        "format": "int64",
                        "type": "integer"
                    },
                    "name": {
                        "example": "Widget",
                        "type": "string"
                    }
                },
                "required": [
                    "description",
                    "id",
                    "name"
                ],
                "type": "object"
            },
            "api.Problem": {
                "properties": {
                    "detail": {
                        "example": "Item 9 was not found",
                        "type": "string"
                    },
                    "status": {
                        "example": 404,
                        "type": "integer"
                    },
                    "title": {
                        "example": "Item not found",
                        "type": "string"
                    },
                    "type": {
                        "example": "about:blank",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "info": {
        "description": "Item CRUD service for the simpsonm09 repository fixture",
        "title": "Simpsonm09 Fixture Go API",
        "version": "0.1.0"
    },
    "openapi": "3.1.0",
    "paths": {
        "/items": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/api.ItemResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "List every item",
                "tags": [
                    "Items"
                ]
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.ItemRequest",
                                "description": "Item to create",
                                "summary": "request"
                            }
                        }
                    },
                    "description": "Item to create",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ItemResponse"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.Problem"
                                }
                            }
                        },
                        "description": "Validation failed"
                    }
                },
                "summary": "Create an item",
                "tags": [
                    "Items"
                ]
            }
        },
        "/items/{id}": {
            "delete": {
                "parameters": [
                    {
                        "description": "Item id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "int64",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Item deleted"
                    },
                    "404": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.Problem"
                                }
                            }
                        },
                        "description": "Item not found"
                    }
                },
                "summary": "Delete an item",
                "tags": [
                    "Items"
                ]
            },
            "get": {
                "parameters": [
                    {
                        "description": "Item id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "int64",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ItemResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.Problem"
                                }
                            }
                        },
                        "description": "Item not found"
                    }
                },
                "summary": "Get one item by id",
                "tags": [
                    "Items"
                ]
            },
            "put": {
                "parameters": [
                    {
                        "description": "Item id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "int64",
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.ItemRequest",
                                "description": "Item to store",
                                "summary": "request"
                            }
                        }
                    },
                    "description": "Item to store",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ItemResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.Problem"
                                }
                            }
                        },
                        "description": "Validation failed"
                    },
                    "404": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.Problem"
                                }
                            }
                        },
                        "description": "Item not found"
                    }
                },
                "summary": "Replace an item",
                "tags": [
                    "Items"
                ]
            }
        }
    }
}